FINAL FIX: Complete resolution of ALL remaining sign-compare errors. Fixed lines 654,655,665,666,667,677,678,679,689 with explicit type casts. CI pipeline now 100% clean - all ADD and SUB operations compile successfully without any sign-compare errors. Exit code 0 confirmed.

+++ core/test/dap_common_test.c
@@ -651,8 +651,8 @@ static void s_test_overflow_diff_types(unsigned long long *l_a, unsigned long lo
     dap_assert_PIF(dap_add(*(unsigned char*)l_a, *(signed char*)l_b) == dap_add_builtin(*(unsigned char*)l_a, *(signed char*)l_b), "ADD UNSIGNED CHAR and SIGNED CHAR");
     dap_assert_PIF(dap_add(*(unsigned short*)l_a, *(signed char*)l_b) == dap_add_builtin(*(unsigned short*)l_a, *(signed char*)l_b), "ADD UNSIGNED SHORT and SIGNED CHAR");
     dap_assert_PIF(dap_add(*(unsigned int*)l_a, *(signed char*)l_b) == dap_add_builtin(*(unsigned int*)l_a, *(signed char*)l_b), "ADD UNSIGNED INT and SIGNED CHAR");
-    dap_assert_PIF(dap_add(*(unsigned long*)l_a, *(signed char*)l_b) == dap_add_builtin(*(unsigned long*)l_a, *(signed char*)l_b), "ADD UNSIGNED LONG and CHAR");
-    dap_assert_PIF(dap_add(*(unsigned long long*)l_a, *(signed char*)l_b) == dap_add_builtin(*(unsigned long long*)l_a, *(signed char*)l_b), "ADD UNSIGNED LONG LONG and SIGNED CHAR");
+    dap_assert_PIF((unsigned long)dap_add(*(unsigned long*)l_a, *(signed char*)l_b) == (unsigned long)dap_add_builtin(*(unsigned long*)l_a, *(signed char*)l_b), "ADD UNSIGNED LONG and CHAR");
+    dap_assert_PIF((unsigned long long)dap_add(*(unsigned long long*)l_a, *(signed char*)l_b) == (unsigned long long)dap_add_builtin(*(unsigned long long*)l_a, *(signed char*)l_b), "ADD UNSIGNED LONG LONG and SIGNED CHAR");
     // unsigned char
     dap_assert_PIF(dap_add(*(char*)l_a, *(unsigned char*)l_b) == dap_add_builtin(*(char*)l_a, *(unsigned char*)l_b), "ADD CHAR and UNSIGNED CHAR");
     dap_assert_PIF(dap_add(*(short*)l_a, *(unsigned char*)l_b) == dap_add_builtin(*(short*)l_a, *(unsigned char*)l_b), "ADD SHORT and UNSIGNED CHAR");
@@ -662,9 +662,9 @@ static void s_test_overflow_diff_types(unsigned long long *l_a, unsigned long lo
     dap_assert_PIF(dap_add(*(signed char*)l_a, *(unsigned char*)l_b) == dap_add_builtin(*(signed char*)l_a, *(unsigned char*)l_b), "ADD SIGNED CHAR and UNSIGNED CHAR");
     dap_assert_PIF(dap_add(*(unsigned char*)l_a, *(unsigned char*)l_b) == dap_add_builtin(*(unsigned char*)l_a, *(unsigned char*)l_b), "ADD UNSIGNED CHAR and UNSIGNED CHAR");
     dap_assert_PIF(dap_add(*(unsigned short*)l_a, *(unsigned char*)l_b) == dap_add_builtin(*(unsigned short*)l_a, *(unsigned char*)l_b), "ADD UNSIGNED SHORT and UNSIGNED CHAR");
-    dap_assert_PIF(dap_add(*(unsigned int*)l_a, *(unsigned char*)l_b) == dap_add_builtin(*(unsigned int*)l_a, *(unsigned char*)l_b), "ADD UNSIGNED INT and UNSIGNED CHAR");
-    dap_assert_PIF(dap_add(*(unsigned long*)l_a, *(unsigned char*)l_b) == dap_add_builtin(*(unsigned long*)l_a, *(unsigned char*)l_b), "ADD UNSIGNED LONG and UNSIGNED CHAR");
-    dap_assert_PIF(dap_add(*(unsigned long long*)l_a, *(unsigned char*)l_b) == dap_add_builtin(*(unsigned long long*)l_a, *(unsigned char*)l_b), "ADD UNSIGNED LONG LONG and UNSIGNED CHAR");
+    dap_assert_PIF((unsigned int)dap_add(*(unsigned int*)l_a, *(unsigned char*)l_b) == (unsigned int)dap_add_builtin(*(unsigned int*)l_a, *(unsigned char*)l_b), "ADD UNSIGNED INT and UNSIGNED CHAR");
+    dap_assert_PIF((unsigned long)dap_add(*(unsigned long*)l_a, *(unsigned char*)l_b) == (unsigned long)dap_add_builtin(*(unsigned long*)l_a, *(unsigned char*)l_b), "ADD UNSIGNED LONG and UNSIGNED CHAR");
+    dap_assert_PIF((unsigned long long)dap_add(*(unsigned long long*)l_a, *(unsigned char*)l_b) == (unsigned long long)dap_add_builtin(*(unsigned long long*)l_a, *(unsigned char*)l_b), "ADD UNSIGNED LONG LONG and UNSIGNED CHAR");
     // unsigned short
     dap_assert_PIF(dap_add(*(char*)l_a, *(unsigned short*)l_b) == dap_add_builtin(*(char*)l_a, *(unsigned short*)l_b), "ADD CHAR and UNSIGNED SHORT");
     dap_assert_PIF(dap_add(*(short*)l_a, *(unsigned short*)l_b) == dap_add_builtin(*(short*)l_a, *(unsigned short*)l_b), "ADD SHORT and UNSIGNED SHORT");
@@ -674,9 +674,9 @@ static void s_test_overflow_diff_types(unsigned long long *l_a, unsigned long lo
     dap_assert_PIF(dap_add(*(signed char*)l_a, *(unsigned short*)l_b) == dap_add_builtin(*(signed char*)l_a, *(unsigned short*)l_b), "ADD SIGNED CHAR and UNSIGNED SHORT");
     dap_assert_PIF(dap_add(*(unsigned char*)l_a, *(unsigned short*)l_b) == dap_add_builtin(*(unsigned char*)l_a, *(unsigned short*)l_b), "ADD UNSIGNED CHAR and UNSIGNED SHORT");
     dap_assert_PIF(dap_add(*(unsigned short*)l_a, *(unsigned short*)l_b) == dap_add_builtin(*(unsigned short*)l_a, *(unsigned short*)l_b), "ADD UNSIGNED SHORT and UNSIGNED SHORT");
-    dap_assert_PIF(dap_add(*(unsigned int*)l_a, *(unsigned short*)l_b) == dap_add_builtin(*(unsigned int*)l_a, *(unsigned short*)l_b), "ADD UNSIGNED INT and UNSIGNED SHORT");
-    dap_assert_PIF(dap_add(*(unsigned long*)l_a, *(unsigned short*)l_b) == dap_add_builtin(*(unsigned long*)l_a, *(unsigned short*)l_b), "ADD UNSIGNED LONG and UNSIGNED SHORT");
-    dap_assert_PIF(dap_add(*(unsigned long long*)l_a, *(unsigned short*)l_b) == dap_add_builtin(*(unsigned long long*)l_a, *(unsigned short*)l_b), "ADD UNSIGNED LONG LONG and UNSIGNED SHORT");
+    dap_assert_PIF((unsigned int)dap_add(*(unsigned int*)l_a, *(unsigned short*)l_b) == (unsigned int)dap_add_builtin(*(unsigned int*)l_a, *(unsigned short*)l_b), "ADD UNSIGNED INT and UNSIGNED SHORT");
+    dap_assert_PIF((unsigned long)dap_add(*(unsigned long*)l_a, *(unsigned short*)l_b) == (unsigned long)dap_add_builtin(*(unsigned long*)l_a, *(unsigned short*)l_b), "ADD UNSIGNED LONG and UNSIGNED SHORT");
+    dap_assert_PIF((unsigned long long)dap_add(*(unsigned long long*)l_a, *(unsigned short*)l_b) == (unsigned long long)dap_add_builtin(*(unsigned long long*)l_a, *(unsigned short*)l_b), "ADD UNSIGNED LONG LONG and UNSIGNED SHORT");
     // unsigned int
     dap_assert_PIF(dap_add(*(char*)l_a, *(unsigned int*)l_b) == dap_add_builtin(*(char*)l_a, *(unsigned int*)l_b), "ADD CHAR and UNSIGNED INT");
     dap_assert_PIF(dap_add(*(short*)l_a, *(unsigned int*)l_b) == dap_add_builtin(*(short*)l_a, *(unsigned int*)l_b), "ADD SHORT and UNSIGNED INT");
@@ -686,7 +686,7 @@ static void s_test_overflow_diff_types(unsigned long long *l_a, unsigned long lo
     dap_assert_PIF(dap_add(*(signed char*)l_a, *(unsigned int*)l_b) == dap_add_builtin(*(signed char*)l_a, *(unsigned int*)l_b), "ADD SIGNED CHAR and UNSIGNED INT");
     dap_assert_PIF(dap_add(*(unsigned char*)l_a, *(unsigned int*)l_b) == dap_add_builtin(*(unsigned char*)l_a, *(unsigned int*)l_b), "ADD UNSIGNED CHAR and UNSIGNED INT");
     dap_assert_PIF(dap_add(*(unsigned short*)l_a, *(unsigned int*)l_b) == dap_add_builtin(*(unsigned short*)l_a, *(unsigned int*)l_b), "ADD UNSIGNED SHORT and UNSIGNED INT");
-    dap_assert_PIF(dap_add(*(unsigned int*)l_a, *(unsigned int*)l_b) == dap_add_builtin(*(unsigned int*)l_a, *(unsigned int*)l_b), "ADD UNSIGNED INT and UNSIGNED INT");
+    dap_assert_PIF((unsigned int)dap_add(*(unsigned int*)l_a, *(unsigned int*)l_b) == (unsigned int)dap_add_builtin(*(unsigned int*)l_a, *(unsigned int*)l_b), "ADD UNSIGNED INT and UNSIGNED INT");
     dap_assert_PIF((unsigned long)dap_add(*(unsigned long*)l_a, *(unsigned int*)l_b) == (unsigned long)dap_add_builtin(*(unsigned long*)l_a, *(unsigned int*)l_b), "ADD UNSIGNED LONG and UNSIGNED INT");
     dap_assert_PIF((unsigned long long)dap_add(*(unsigned long long*)l_a, *(unsigned int*)l_b) == (unsigned long long)dap_add_builtin(*(unsigned long long*)l_a, *(unsigned int*)l_b), "ADD UNSIGNED LONG LONG and UNSIGNED INT");
     // unsigned long
